changed config class, parsing functions, filenames
After learning there could be any config key value pair, a change was made to stop looking for specific keys and take in all keys given with their respective values. Better error checking was also added for Booleans and Numbers to handle if numbers have letters in the string such as "28934a", and to read in lowercased Bool values like "Yes".